segmentation-engine/requirements.txt
albumentations==1.1.0
antlr4-python3-runtime==4.8
certifi==2021.5.30
cycler==0.10.0
efficientnet-pytorch==0.6.3
easing-functions==1.0.4
fvcore==0.1.5.post20210630
iopath==0.1.9
joblib==1.0.1
kiwisolver==1.3.1
matplotlib==3.4.2
munch==2.5.0
numpy==1.20.0
olefile==0.46
omegaconf==2.1.0
opencv-python==4.5.3.56
Pillow==6.2.2
portalocker==2.3.0
prefetch-generator==1.0.1
pretrainedmodels==0.7.4
pyparsing==2.4.7
python-dateutil==2.8.2
PyYAML==5.4.1
scikit-learn==0.24.2
scipy==1.7.0
segmentation_models_pytorch==0.2.0
sklearn==0.0
tabulate==0.8.9
termcolor==1.1.0
threadpoolctl==2.2.0
timm==0.4.12
torchvision==0.8.2
torchstat==0.0.7
tqdm==4.61.2
yacs==0.1.8
